Implement 64-bit syscall return values and fix sys_lseek() arguments bitness

+++ kern/include/syscall.h
@@ -65,7 +65,7 @@ int sys_dup2(int oldfd, int newfd, int *retval);
 int sys_close(int fd);
 int sys_read(int fd, userptr_t buf, size_t nbytes, int *retval);
 int sys_write(int fd, userptr_t buf, size_t nbytes, int *retval);
-int sys_lseek(int fd, off_t offset, int whence, int *retval);
+int sys_lseek(int fd, off_t offset, int whence, int64_t *retval);
 int sys_chdir(userptr_t pathname);
 int sys___getcwd(userptr_t buf, size_t buflen, int *retval);
